refactored StallGuard - fixed demo scripts - moved StallGuard code into own mixin class - fixed StallGuard on TMC2240 (diag0_pushpull and diag0_stall needed to be activated) - renamed sgresult and sgthrs reg values in order to have them consistend between drivers
