+++ .vscode/settings.json
@@ -4,5 +4,6 @@
         "--disable=W0511",
         "--disable=W0012",
         "--extension-pkg-whitelist=RPi",
-    ]
+    ],
+    "files.trimTrailingWhitespace": true
 }
