+++ src/tmc_driver/tmc_2209.py
@@ -10,16 +10,13 @@
 import statistics
 import types
 from .tmc_220x import *
+from ._tmc_stallguard import StallGuard
 from ._tmc_gpio_board import GpioPUD
 from .reg._tmc2209_reg import *
 
 
-class Tmc2209(Tmc220x):
+class Tmc2209(Tmc220x, StallGuard):
     """Tmc2209"""
-
-    _pin_stallguard:int = None
-    _sg_callback:types.FunctionType = None
-    _sg_threshold:int = 100             # threshold for stallguard
 
 
 
@@ -78,6 +75,7 @@ class Tmc2209(Tmc220x):
         self.tmc_logger.log("TMC2209 Init finished", Loglevel.INFO)
 
 
+
     def __del__(self):
         """destructor"""
         if self._deinit_finished is False:
@@ -89,49 +87,8 @@ class Tmc2209(Tmc220x):
 
 
 
-    def set_stallguard_callback(self, pin_stallguard, threshold, callback,
-                                min_speed = 100):
-        """set a function to call back, when the driver detects a stall
-        via stallguard
-        high value on the diag pin can also mean a driver error
-
-        Args:
-            pin_stallguard (int): pin needs to be connected to DIAG
-            threshold (int): value for SGTHRS
-            callback (func): will be called on StallGuard trigger
-            min_speed (int): min speed [steps/s] for StallGuard (Default value = 100)
-        """
-        self.tmc_logger.log(f"setup stallguard callback on GPIO {pin_stallguard}", Loglevel.INFO)
-        self.tmc_logger.log(f"StallGuard Threshold: {threshold} | minimum Speed: {min_speed}", Loglevel.INFO)
-
-        self._set_stallguard_threshold(threshold)
-        self._set_coolstep_threshold(tmc_math.steps_to_tstep(min_speed, self.get_microstepping_resolution()))
-        self._sg_callback = callback
-        self._pin_stallguard = pin_stallguard
-
-        tmc_gpio.gpio_setup(self._pin_stallguard, GpioMode.IN, pull_up_down=GpioPUD.PUD_DOWN)
-        # first remove existing events
-        tmc_gpio.gpio_remove_event_detect(self._pin_stallguard)
-        tmc_gpio.gpio_add_event_detect(self._pin_stallguard, self.stallguard_callback)
-
-
-
-    def stallguard_callback(self, gpio_pin):
-        """the callback function for StallGuard.
-        only checks whether the duration of the current movement is longer than
-        _sg_delay and then calls the actual callback
-
-        Args:
-            gpio_pin (int): pin number of the interrupt pin
-        """
-        del gpio_pin
-        if self._sg_callback is None:
-            self.tmc_logger.log("StallGuard callback is None", Loglevel.DEBUG)
-            return
-        self._sg_callback()
-
-
-
+# Tmc2209 methods
+# ----------------------------
     def do_homing(self, diag_pin, revolutions = 10, threshold = None, speed_rpm = None) -> bool:
         """homes the motor in the given direction using stallguard.
         this method is using vactual to move the motor and an interrupt on the DIAG pin
@@ -197,7 +154,7 @@ class Tmc2209(Tmc220x):
         if not isinstance(self.tmc_mc, TmcMotionControlStepDir):
             self.tmc_logger.log("do_homing2 only works with STEP/DIR Control", Loglevel.ERROR)
             return
-        sg_results = []
+        sgresults = []
 
         if threshold is not None:
             self._sg_threshold = threshold
@@ -231,88 +188,29 @@ class Tmc2209(Tmc220x):
             if self.tmc_mc.run_speed(): #returns true, when a step is made
                 step_counter += 1
                 self.tmc_mc.compute_new_speed()
-                sg_result = self.get_stallguard_result()
-                sg_results.append(sg_result)
-                if len(sg_results)>20:
-                    sg_result_average = statistics.mean(sg_results[-6:])
-                    if sg_result_average < self._sg_threshold:
+                sgresult = self.get_stallguard_result()
+                sgresults.append(sgresult)
+                if len(sgresults)>20:
+                    sgresult_average = statistics.mean(sgresults[-6:])
+                    if sgresult_average < self._sg_threshold:
                         break
 
         if step_counter<self.tmc_mc.steps_per_rev:
             self.tmc_logger.log("homing successful",Loglevel.INFO)
             self.tmc_logger.log(f"Stepcounter: {step_counter}",Loglevel.DEBUG)
-            self.tmc_logger.log(str(sg_results),Loglevel.DEBUG)
+            self.tmc_logger.log(str(sgresults),Loglevel.DEBUG)
             self.tmc_mc.current_pos = 0
         else:
             self.tmc_logger.log("homing failed", Loglevel.INFO)
             self.tmc_logger.log(f"Stepcounter: {step_counter}", Loglevel.DEBUG)
-            self.tmc_logger.log(str(sg_results),Loglevel.DEBUG)
+            self.tmc_logger.log(str(sgresults),Loglevel.DEBUG)
 
         self.tmc_logger.log("---", Loglevel.INFO)
 
 
 
-    def get_stallguard_result(self):
-        """return the current stallguard result
-        its will be calculated with every fullstep
-        higher values means a lower motor load
-
-        Returns:
-            sg_result (int): StallGuard Result
-        """
-        self.sgresult.read()
-        return self.sgresult.sgresult
-
-
-
-    def _set_stallguard_threshold(self, threshold):
-        """sets the register bit "SGTHRS" to to a given value
-        this is needed for the stallguard interrupt callback
-        SG_RESULT becomes compared to the double of this threshold.
-        SG_RESULT ≤ SGTHRS*2
-
-        Args:
-            threshold (int): value for SGTHRS
-        """
-        self.sgthrs.modify("sgthrs", threshold)
-
-
-
-    def _set_coolstep_threshold(self, threshold):
-        """This  is  the  lower  threshold  velocity  for  switching
-        on  smart energy CoolStep and StallGuard to DIAG output. (unsigned)
-
-        Args:
-            threshold (int): threshold velocity for coolstep
-        """
-        self.tcoolthrs.modify("tcoolthrs", threshold)
-
-
-
-    def enable_coolstep(self, semin_sg:int = 150, semax_sg:int = 200, seup:int = 1, sedn:int = 3, min_speed:int = 100):
-        """enables coolstep and sets the parameters for coolstep
-        The values for semin etc. can be tested with the test_stallguard_threshold function
-
-        Args:
-            semin_sg (int): lower threshold. Current will be increased if SG_Result goes below this
-            semax_sg (int): upper threshold. Current will be decreased if SG_Result goes above this
-            seup (int): current increment step
-            sedn (int): number of SG_Result readings for each current decrement
-        """
-        semax_sg = semax_sg - semin_sg
-
-        self.coolconf.read()
-        self.coolconf.semin = round(max(0, min(semin_sg/32, 15)))
-        self.coolconf.semax = round(max(0, min(semax_sg/32, 15)))
-        self.coolconf.seimin = 1        # scale down to until 1/4 of IRun (7 - 31)
-        self.coolconf.seup = seup
-        self.coolconf.sedn = sedn
-        self.coolconf.write_check()
-
-        self._set_coolstep_threshold(tmc_math.steps_to_tstep(min_speed, self.get_microstepping_resolution()))
-
-
-
+# Test methods
+# ----------------------------
     def test_stallguard_threshold(self, steps):
         """test method for tuning stallguard threshold
 
@@ -322,7 +220,6 @@ class Tmc2209(Tmc220x):
         Args:
             steps (int):
         """
-
         self.tmc_logger.log("---", Loglevel.INFO)
         self.tmc_logger.log("test_stallguard_threshold", Loglevel.INFO)
 
@@ -339,12 +236,8 @@ class Tmc2209(Tmc220x):
             stallguard_result = self.get_stallguard_result()
             self.drvstatus.read()
             cs_actual = self.drvstatus.cs_actual
-            # stallguard_result = self.get_stallguard_result()
 
             self.tmc_logger.log(f"{self.tmc_mc.movement_phase} | {stallguard_result} | {cs_actual}",
-                        Loglevel.INFO)
-
-            self.tmc_logger.log(f"{self.tmc_mc.movement_phase} | {stallguard_result}",
                         Loglevel.INFO)
 
             if (self.tmc_mc.movement_phase == MovementPhase.ACCELERATING and
